Move the Post page to a child folder.
