+++ BlogMaui/Areas/Blog/Posts/PostEditViewModel.cs
@@ -1,4 +1,4 @@
-﻿using BlogMaui.Areas.Blog;
+﻿using BlogMaui.Areas.Blog.Posts;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using Jinaga;
