+++ BlogMaui/Areas/Blog/Posts/PostListPage.xaml
@@ -2,7 +2,7 @@
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:root="clr-namespace:BlogMaui"
-             x:Class="BlogMaui.Areas.Blog.PostListPage"
+             x:Class="BlogMaui.Areas.Blog.Posts.PostListPage"
              Title="Posts">
 
     <Grid RowDefinitions="Auto,*,Auto,Auto">
