+++ BlogMaui/Areas/Blog/Posts/PostListPage.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Extensions.Logging;
 
-namespace BlogMaui.Areas.Blog;
+namespace BlogMaui.Areas.Blog.Posts;
 
 // The refresh view does not use the command to indicate
 // that the user has initiated a refresh. It executes the
