+++ BlogMaui/Areas/Blog/Posts/PostPage.xaml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="BlogMaui.Areas.Blog.PostPage"
+             x:Class="BlogMaui.Areas.Blog.Posts.PostPage"
              xmlns:fonts="clr-namespace:BlogMaui.Fonts"
              Title="Post">
     <Grid RowDefinitions="Auto,*" ColumnDefinitions="*">
