+++ BlogMaui/Areas/Blog/Posts/PostPage.xaml.cs
@@ -1,4 +1,4 @@
-namespace BlogMaui.Areas.Blog;
+namespace BlogMaui.Areas.Blog.Posts;
 
 public partial class PostPage : ContentPage
 {
