+++ BlogMaui/Areas/Blog/Posts/PostViewModel.cs
@@ -4,7 +4,7 @@ using Jinaga;
 using System.Collections.Immutable;
 using System.Windows.Input;
 
-namespace BlogMaui.Areas.Blog;
+namespace BlogMaui.Areas.Blog.Posts;
 
 public partial class PostViewModel : ObservableObject, IQueryAttributable
 {
